#178 - Windows support with PowerShell added
